refactor: reorganize config handlers with ABC and modular structure
Move config handlers from single monolithic file into their respective
task packages while maintaining backward compatibility via lazy imports.
Add ConfigHandler abstract base class to enforce consistent interface.

Changes:
- Add ConfigHandler ABC requiring run() method and ids attribute
- Move config handlers to task packages:
  * DataConfigHandler → data/data_config_handler.py
  * TrainingConfigHandler → training/training_config_handler.py
  * MetricsConfigHandler → metrics/metrics_config_handler.py
  * PlottingConfigHandler → plotting/plotting_config_handler.py
- Refactor configs_handler.py to registry-only (ABC + helper functions)
- Add lazy imports via __getattr__ for backward compatibility
- Register handlers automatically in wf_psf/__init__.py

Tests:
- Add TestConfigHandlerABC for ABC enforcement
- Reorganize into TestConfigRegistry and TestTrainingConfigHandler classes
- Add tests for all registered handlers inheriting from ConfigHandler
- Update import paths throughout test suite

Backward compatibility:
- Old imports still work via __getattr__ lazy loading
- Registry functions (get_run_config, set_run_config) unchanged
- TYPE_CHECKING guard satisfies static analysis tools
