+++ changelog.d/20260225_104813_jennifer.pollack_refactor_config_handler.md
@@ -1,0 +1,42 @@
+<!--
+A new scriv changelog fragment.
+
+Uncomment the section that is right (remove the HTML comment wrapper).
+For top level release notes, leave all the headers commented out.
+-->
+
+
+### Breaking changes
+
+- Config handlers moved to their respective task packages. Import paths changed:
+  - `TrainingConfigHandler`: now in `wf_psf.training.training_config_handler`
+  - `MetricsConfigHandler`: now in `wf_psf.metrics.metrics_config_handler`
+  - `PlottingConfigHandler`: now in `wf_psf.plotting.plotting_config_handler`
+  - `DataConfigHandler`: now in `wf_psf.data.data_config_handler`
+  - Backward compatibility maintained via `wf_psf.utils.configs_handler`
+
+
+### New features
+
+- Added `ConfigHandler` abstract base class for type safety and enforced interface compliance
+- All config handlers now inherit from `ConfigHandler` ABC and must implement `run()` method
+- Config handler registry pattern enhanced with ABC enforcement
+
+### Bug fixes
+
+- A bullet item for the Bug fixes category.
+
+-->
+<!--
+### Performance improvements
+
+- A bullet item for the Performance improvements category.
+
+-->
+
+### Internal changes
+
+- Reorganized config handlers into modular structure with each handler in its task package
+- Improved config handler registration system with abstract base class
+- Enhanced test coverage for config handler ABC enforcement and registry dispatch
+- Registry imports handlers automatically when `wf_psf` package is imported
